Update odd or even.py

+++ Beginner/odd or even.py
@@ -1,4 +1,4 @@
-n=float(input())
+n=int(input())
 if(n%2==0):
 	print("even")
 else:
